+++ src/models/index.ts
@@ -1,6 +1,6 @@
 import { bootModels } from 'soukai';
 
-import type { JsonLD } from '@/solid/utils/RDF';
+import type { JsonLD, JsonLDResource } from '@/solid/utils/RDF';
 
 import type DeletesModels from './mixins/DeletesModels';
 import type { RDFContext } from './mixins/SerializesToJsonLD';
@@ -8,6 +8,7 @@ import type SerializesToJsonLD from './mixins/SerializesToJsonLD';
 
 import SolidContainerModel from './SolidContainerModel';
 import SolidDocument from './SolidDocument';
+import SolidModelMetadata from './SolidModelMetadata';
 
 export * from './inference';
 export * from './relations/index';
@@ -16,11 +17,13 @@ export * from './SolidModel';
 export {
     SolidContainerModel,
     SolidDocument,
+    SolidModelMetadata,
 };
 
 export type {
     DeletesModels,
     JsonLD,
+    JsonLDResource,
     RDFContext,
     SerializesToJsonLD,
 };
@@ -31,5 +34,5 @@ export type {
 } from './fields';
 
 export function bootSolidModels(): void {
-    bootModels({ SolidDocument });
+    bootModels({ SolidDocument, SolidModelMetadata });
 }