+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Upgraded tooling.
 - Replaced `jsonld-streaming-parser` and `jsonld-streaming-serializer` dependencies with `jsonld`.
 - Make sure to check Soukai's changelog.
+- Automatic timestamps have been refactored into a metadata resource.
 
 ## [v0.4.2](https://github.com/NoelDeMartin/soukai-solid/releases/tag/v0.4.2) - 2021-03-28